Conformance controller modified
