+++ OgcApi.Net/Controllers/ConformanceController.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Options;
 using OgcApi.Net.Options;
 using OgcApi.Net.Resources;
 using System;
+using System.Linq;
 using System.Collections.Generic;
 
 namespace OgcApi.Net.Controllers
@@ -14,13 +15,13 @@ namespace OgcApi.Net.Controllers
     [ApiExplorerSettings(GroupName = "ogc")]
     public class ConformanceController : ControllerBase
     {
-        private readonly ConformanceOptions _apiOptions;
+        private readonly OgcApiOptions _apiOptions;
 
         private readonly ILogger _logger;
 
         public ConformanceController(IOptionsMonitor<OgcApiOptions> apiOptions, ILoggerFactory logger)
         {
-            _apiOptions = apiOptions.CurrentValue.Conformance;
+            _apiOptions = apiOptions.CurrentValue;
 
             _logger = logger.CreateLogger("OGC API Features ConformanceController");
 
@@ -45,24 +46,35 @@ namespace OgcApi.Net.Controllers
         {
             _logger.LogTrace($"Get conformance with parameters {Request.QueryString}");
 
-            if (_apiOptions?.ConformsTo == null || _apiOptions.ConformsTo.Count == 0)
+            if (_apiOptions?.Conformance?.ConformsTo == null || _apiOptions.Conformance.ConformsTo.Count == 0)
             {
+                var conformsTo = new List<Uri>();
+                
+                if (_apiOptions.Collections.Items.Any(item => item.Features != null))
+                {
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core"));
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30"));
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/html"));
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson"));
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-features-2/1.0/conf/crs"));
+                }
+
+                if (_apiOptions.Collections.Items.Any(item => item.Tiles != null))
+                {
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/tileset"));
+                    conformsTo.Add(new("http://www.opengis.net/spec/ogcapi-tiles-1/1.0/conf/tilesets-list"));
+                }
+
+
                 return new Conformance
                 {
-                    ConformsTo = new List<Uri>
-                    {
-                        new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core"),
-                        new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30"),
-                        new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/html"),
-                        new("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson"),
-                        new("http://www.opengis.net/spec/ogcapi-features-2/1.0/conf/crs")
-                    }
+                    ConformsTo = conformsTo
                 };
             }
 
             return new Conformance
             {
-                ConformsTo = _apiOptions.ConformsTo
+                ConformsTo = _apiOptions.Conformance.ConformsTo
             };
         }
     }
